Unitu/users/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2
flask-migrate
flask-bcrypt
pyjwt
requests
Flask-Testing
mock
coverage
parameterized
flask-cors
